+++ Domain/Entities/Product.cs
@@ -4,7 +4,7 @@ namespace Domain.Entities;
 
 public class Product
 {
-    public ulong Id { get; set; }
+    public long Id { get; set; }
     public Guid Uuid { get; set; }
     public string? IgId { get; set; }
     public required string Name { get; set; }
@@ -12,15 +12,15 @@ public class Product
     public decimal Price { get; set; }
     public decimal OriginalPrice { get; set; }
     public byte DiscountPercentage { get; set; }
-    public uint Stock { get; set; }
+    public long Stock { get; set; }
     public string? IgCode { get; set; }
     public string? IgThumbnailSrc { get; set; }
     public string? IgDisplayUrl { get; set; }
-    public uint? IgLikeCount { get; set; }
-    public uint? IgCommentCount { get; set; }
+    public long? IgLikeCount { get; set; }
+    public long? IgCommentCount { get; set; }
     public byte? IgCarouselMediaCount { get; set; }
     public string? IgVideoUrl { get; set; }
-    public uint ShopId { get; set; }
+    public long ShopId { get; set; }
     public bool HasDiscount { get; set; }
     public bool IsAvailable { get; set; }
     public bool IsNew { get; set; }
@@ -43,8 +43,8 @@ public class Product
 
 public class Dimensions
 {
-    public ushort? Hieght { get; set; }
-    public ushort? Width { get; set; }
+    public int? Hieght { get; set; }
+    public int? Width { get; set; }
 }
 
 public class Caption
