+++ Domain/Entities/User.cs
@@ -5,7 +5,7 @@ namespace Domain.Entities;
 
 public class User
 {
-    public uint Id { get; set; }
+    public long Id { get; set; }
     public Guid Uuid { get; set; }
     public string? Email { get; set; }
     public string? PhoneNumber { get; set; }
@@ -13,8 +13,8 @@ public class User
     public string? LastName { get; set; }
     public DateOnly? BirthDate { get; set; }
     public string? Password { get; set; }
-    public ushort? LastOtp { get; set; }
-    public uint? ShopId { get; set; }
+    public int? LastOtp { get; set; }
+    public long? ShopId { get; set; }
     public UserRole Role { get; set; }
     public bool IsDeleted { get; set; }
     public DateTime CreateDate { get; set; }
