+++ Infrastructure/Dtos/InstagramDto.cs
@@ -11,8 +11,8 @@ public class InstagramDto
         public string? ThumbnailSrc { get; set; }
         public string? DisplayUrl { get; set; }
         public Dimensions? Dimensions { get; set; }
-        public uint? LikeCount { get; set; }
-        public uint? CommentCount { get; set; }
+        public long? LikeCount { get; set; }
+        public long? CommentCount { get; set; }
         public byte? CarouselMediaCount { get; set; }
         public string? VideoUrl { get; set; }
         public bool? IsVideo { get; set; }
@@ -28,7 +28,7 @@ public class InstagramDto
         public string? Username { get; set; }
         public string? FullName { get; set; }
         public string? ProfilePictureUrl { get; set; }
-        public uint? Followers { get; set; }
+        public long? Followers { get; set; }
         public bool? IsVerified { get; set; }
     }
