+++ Infrastructure/Interfaces/IUserRepository.cs
@@ -5,14 +5,14 @@ namespace Application.Interfaces;
 
 public interface IUserRepository
 {
-    Task<uint> CreateTemporaryUser(string phoneNumber);
-    Task<ushort> SetLastOtp(uint userId);
+    Task<long> CreateTemporaryUser(string phoneNumber);
+    Task<int> SetLastOtp(long userId);
     Task<User> RegisterUser(UserRegisterDto userRegisterDto);
-    Task<uint> SetNewPasswordForUser(UserCredentialsDto userCredentialsDto);
-    Task<User> GetUserById(uint userId);
+    Task<long> SetNewPasswordForUser(UserCredentialsDto userCredentialsDto);
+    Task<User> GetUserById(long userId);
     Task<User> GetUserByPhoneNumber(string phoneNumber);
     Task<User> GetUserByCredentials(UserCredentialsDto userGetByCredentialsDto);
-    Task<User> GetShopOwnerByShopId(uint shopId);
+    Task<User> GetShopOwnerByShopId(long shopId);
     Task<List<User>> GetAllAdmins();
-    Task<uint> DeleteUser(uint userId);
+    Task<long> DeleteUser(long userId);
 }
